Back_end/requirements.txt
fastapi
uvicorn
sqlalchemy
groq
python-dotenv
axios
pymysql
